added admin role and related features

+++ backend/src/controllers/storeController.ts
@@ -14,22 +14,34 @@ const findById = async (req: Request, res: Response) => {
 
 const create = async (req: Request, res: Response) => {
   const data = req.body;
-  const store = await storeService.create(data);
-  res.status(201).json(store);
+  if (id === req.user.id || req.user.role === "admin") {
+    const store = await storeService.create(data);
+    res.status(201).json(store);
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 const update = async (req: Request, res: Response) => {
   const data = req.body;
   const id = parseInt(req.params.storeId);
-  await storeService.update({ ...data, id });
-  const store = await storeService.findById(id);
-  res.status(200).json(store);
+  if (id === req.user.id || req.user.role === "admin") {
+    await storeService.update({ ...data, id });
+    const store = await storeService.findById(id);
+    res.status(200).json(store);
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 const destroy = async (req: Request, res: Response) => {
   const id = parseInt(req.params.storeId);
-  await storeService.destroy(id);
-  res.status(200).json({ id: id });
+  if (id === req.user.id || req.user.role === "admin") {
+    await storeService.destroy(id);
+    res.status(200).json({ id: id });
+  } else {
+    res.sendStatus(401);
+  }
 };
 
 export default { getAllStores, findById, create, update, destroy };
